pyproject.toml
[project]
name = "langgraph-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.35.0",
    "langchain-google-genai>=2.1.10",
    "langgraph>=0.6.7",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
